Add files via upload
Pneumonia is a form of acute respiratory infection that affects the lungs. Pneumonia-Detector attempts to automate methods to detect and classify pneumonia from medical x-ray images using a Convolutional Neural Network. It is able to detect correctly 88% of pneumonia cases but it is NOT in any way a substitute for consulting a professional medical examiner.